+++ resources/clj-kondo.exports/sicmutils/sicmutils/hooks/sicmutils/abstract/function.clj
@@ -11,7 +11,10 @@
                 (:children signature))))))
 
 (defn literal-function
-  "from sicmutils.env."
+  "Lints the macro version of `literal-function`, living in `sicmutils.env`.
+
+  If the signature consists of a list-node starting with `->` it's treated as
+  quoted. Else, it's emitted in a vector-node with the `f` entry."
   [{:keys [node]}]
   (let [[_ f sig-or-domain range] (:children node)
         new-node (cond range
@@ -23,11 +26,16 @@
                        :else (api/vector-node [f sig-or-domain]))]
     {:node new-node}))
 
-(defn binding-pair [entry]
-  (cond (api/token-node? entry)
+(defn binding-pair
+  "Given an entry in the binding vector of a call to `with-literal-functions`,
+  returns a 2-vector pair of lintable entries in a `let` form.
+
+  Invalid entries return an empty vector after triggering a linter error."
+  [entry]
+  (cond (and (api/token-node? entry)
+             (simple-symbol? (:value entry)))
         (let [v (api/list-node
-                 [(api/token-node 'quote)
-                  entry])]
+                 [(api/token-node 'quote) entry])]
           [entry v])
 
         (and (or (api/list-node? entry)
@@ -36,17 +44,30 @@
         (let [[sym domain range] (:children entry)]
           [sym (api/vector-node [domain range])])
 
-        :else (throw
-               (ex-info
-                "unknown literal function type, TODO note about required format" {}))))
+        :else
+        (let [{:keys [row col]} (meta entry)]
+          (api/reg-finding!
+           {:message (str "Bindings must be either bare symbols or "
+                          "3-vectors of the form [sym domain range]. "
+                          "Received: "
+                          (pr-str (api/sexpr entry)))
+            :type :sicmutils.abstract.function/invalid-binding
+            :row row
+            :col col})
+          [])))
 
-/(defn with-literal-functions [{:keys [node]}]
-   (let [[_ binding-vec & body] (:children node)
-         bindings (mapcat binding-pair
-                          (:children binding-vec))
-         new-node (api/list-node
-                   (list*
-                    (api/token-node 'let)
-                    (api/vector-node bindings)
-                    body))]
-     {:node new-node}))
+(defn with-literal-functions
+  "Converts a node representing an invocation of
+  the [[sicmutils.abstract.function/with-literal-functions]] macro into a
+  let-style representation of the requested bindings."
+  [{:keys [node]}]
+  (let [[_ binding-vec & body] (:children node)
+        bindings (into []
+                       (mapcat binding-pair)
+                       (:children binding-vec))]
+    {:node
+     (api/list-node
+      (list*
+       (api/token-node 'let)
+       (api/vector-node bindings)
+       body))}))