+++ resources/clj-kondo.exports/sicmutils/sicmutils/hooks/sicmutils/calculus/coordinate.clj
@@ -1,8 +1,8 @@
 (ns hooks.sicmutils.calculus.coordinate
   (:require [clj-kondo.hooks-api :as api]))
 
-(defn ^:no-doc symbols-from-prototype
-  "Generates a list of symbols from the supplied argument prototype. The
+(defn symbols-from-prototype
+  "Generates a list of lintable symbols from the supplied argument prototype. The
   prototype is allowed to be a vector, a list like `(up x y)` or a bare symbol.
   Anything else causes an exception.
 
@@ -12,36 +12,54 @@
              ('#{up down} (:value (first (:children p)))))
         (mapcat symbols-from-prototype (rest (:children p)))
 
-        (api/vector-node? p) (mapcat symbols-from-prototype (:children p))
-        (and (api/token-node? p)
-             (symbol? (:value p)))
-        [p]
-        :else (throw
-               (ex-info (str "Invalid coordinate prototype: " p) {}))))
+        (api/vector-node? p)
+        (mapcat symbols-from-prototype (:children p))
 
-;; from vector-field
-(defn ^:no-doc coordinate-name->vf-name
-  "From the name `n` of a coordinate, produce the name of the coordinate basis
-  vector field (as a symbol)"
+        (and (api/token-node? p) (symbol? (:value p)))
+        [p]
+
+        :else
+        (let [{:keys [row col]} (meta p)]
+          (api/reg-finding!
+           {:message (str "Bindings must be either a vector or list "
+                          "(optionally beginning with `up` or `down`) "
+                          "or a bare symbol. Received: "
+                          (pr-str (api/sexpr p)))
+            :type :sicmutils.calculus.coordinate/invalid-binding
+            :row row
+            :col col})
+          [])))
+
+(defn coordinate-name->vf-name
+  "Given a token-node of coordinate symbolic name, returns a token-node containing
+  the symbolic name of the associated coordinate basis vector field."
   [n]
   (api/token-node
    (symbol (str "d:d" (:value n)))))
 
-;; from form-field
-(defn ^:no-doc coordinate-name->ff-name
-  "From the name of a coordinate, produce the name of the coordinate basis
-  one-form field (as a symbol)"
+(defn coordinate-name->ff-name
+  "Given a token-node of coordinate symbolic name, returns a token-node containing
+  the symbolic name of the associated coordinate basis one-form vector field."
   [n]
   (api/token-node
    (symbol (str \d (:value n)))))
 
-(defn quotify [n]
+(defn quotify
+  "Given some node, returns a 2-vector representing a pair of entries in a
+  let-style binding vector. The value is a quoted node `n`."
+  [n]
   [n (api/list-node [(api/token-node 'quote) n])])
 
-(defn let-coordinates [{:keys [node]}]
+(defn let-coordinates
+  "Converts a node representing an invocation of
+  the [[sicmutils.calculus.coordinate/let-coordinates]] macro into a let-style
+  representation of the requested bindings."
+  [{:keys [node]}]
   (let [[_ binding-vec & body] (:children node)]
     (when-not (even? (count (:children binding-vec)))
-      (throw (ex-info "checking requires an even number of bindings" {})))
+      (throw
+       (ex-info "let-coordinates requires an even number of bindings." {})))
+
     (let [pairs        (partition 2 (:children binding-vec))
           prototypes   (map first pairs)
           c-systems    (map second pairs)
@@ -63,7 +81,11 @@
                      body))]
       {:node new-node})))
 
-(defn using-coordinates [{:keys [node]}]
+(defn using-coordinates
+  "Converts a node representing an invocation of
+  the [[sicmutils.calculus.coordinate/using-coordinates]] macro into a let-style
+  representation of the requested bindings."
+  [{:keys [node]}]
   (let [[sym prototype system & body] (:children node)]
     (let-coordinates
         {:node (api/list-node
@@ -71,12 +93,22 @@
                        (api/vector-node [prototype system])
                        body))})))
 
-(defn ->declare [n]
+(defn ->declare
+  "Given some token-node, returns a 2-vector representing a pair of entries in a
+  let-style binding vector. The value is a quoted node `n`."
+  [n]
   (api/list-node
    [(api/token-node 'declare)
     n]))
 
-(defn define-coordinates [{:keys [node]}]
+(defn define-coordinates
+  "Converts a node representing an invocation of
+  the [[sicmutils.calculus.coordinate/define-coordinates]] macro into a series
+  of declarations for all generated `def` forms, plus a final vector of all
+  entries.
+
+  This last vector will silence any 'not used' warnings."
+  [{:keys [node]}]
   (let [[_ prototype system] (:children node)
         sys-name (api/token-node (symbol (name (:value system))))
         coord-names (symbols-from-prototype prototype)